+++ Cargo.toml
@@ -57,8 +57,8 @@ x25519-dalek = "2.0.0-pre.1"
 [dev-dependencies]
 criterion = { version = "0.4.0", default-features = false, features = ["async_tokio", "plotters", "html_reports"] }
 iai = "0.1.1"
-proptest = "1.0.0"
 permutation = "0.4.1"
+proptest = "1.0.0"
 
 [lib]
 name = "raw_ipa"
