+++ src/protocol/sort/apply_sort/shuffle.rs
@@ -26,7 +26,7 @@ pub trait Resharable<F: Field>: Sized {
         C: Context<F, Share = <Self as Resharable<F>>::Share> + Send;
 }
 
-async fn reshare_objects<F, C, S, T>(input: &[T], ctx: C, to_helper: Role) -> Result<Vec<T>, Error>
+async fn reshare<F, C, S, T>(input: &[T], ctx: C, to_helper: Role) -> Result<Vec<T>, Error>
 where
     C: Context<F, Share = S> + Send,
     F: Field,
@@ -41,11 +41,11 @@ where
     try_join_all(reshares).await
 }
 
-/// `shuffle_or_unshuffle_once` is called for the helpers
+/// `shuffle_once` is called for the helpers
 /// i)   2 helpers receive permutation pair and choose the permutation to be applied
 /// ii)  2 helpers apply the permutation to their shares
 /// iii) reshare to `to_helper`
-async fn shuffle_or_unshuffle_once<F, S, C, I>(
+async fn shuffle_once<F, S, C, I>(
     mut input: Vec<I>,
     random_permutations: (&[u32], &[u32]),
     shuffle_or_unshuffle: ShuffleOrUnshuffle,
@@ -73,18 +73,18 @@ where
             ShuffleOrUnshuffle::Unshuffle => apply(permutation_to_apply, &mut input),
         }
     }
-    reshare_objects(&input, ctx, to_helper).await
+    reshare(&input, ctx, to_helper).await
 }
 
 #[embed_doc_image("shuffle", "images/sort/shuffle.png")]
-/// Shuffle calls `shuffle_or_unshuffle_once` three times with 2 helpers shuffling the shares each time.
-/// Order of calling `shuffle_or_unshuffle_once` is shuffle with (H2, H3), (H3, H1) and (H1, H2).
+/// Shuffle calls `shuffle_once` three times with 2 helpers shuffling the shares each time.
+/// Order of calling `shuffle_once` is shuffle with (H2, H3), (H3, H1) and (H1, H2).
 /// Each shuffle requires communication between helpers to perform reshare.
 /// Infrastructure has a pre-requisite to distinguish each communication step uniquely.
-/// For this, we have three shuffle steps one per `shuffle_or_unshuffle_once` i.e. Step1, Step2 and Step3.
+/// For this, we have three shuffle steps one per `shuffle_once` i.e. Step1, Step2 and Step3.
 /// The Shuffle object receives a step function and appends a `ShuffleStep` to form a concrete step
 /// ![Shuffle steps][shuffle]
-pub async fn shuffle_object_shares<C, F, I, S>(
+pub async fn shuffle_shares<C, F, I, S>(
     input: Vec<I>,
     random_permutations: (&[u32], &[u32]),
     ctx: C,
@@ -95,7 +95,7 @@ where
     I: Resharable<F, Share = S>,
     S: SecretSharing<F>,
 {
-    let input = shuffle_or_unshuffle_once(
+    let input = shuffle_once(
         input,
         random_permutations,
         ShuffleOrUnshuffle::Shuffle,
@@ -103,7 +103,7 @@ where
         Step1,
     )
     .await?;
-    let input = shuffle_or_unshuffle_once(
+    let input = shuffle_once(
         input,
         random_permutations,
         ShuffleOrUnshuffle::Shuffle,
@@ -111,53 +111,12 @@ where
         Step2,
     )
     .await?;
-    shuffle_or_unshuffle_once(
+    shuffle_once(
         input,
         random_permutations,
         ShuffleOrUnshuffle::Shuffle,
         &ctx,
         Step3,
-    )
-    .await
-}
-
-#[embed_doc_image("unshuffle", "images/sort/unshuffle.png")]
-/// Unshuffle calls `shuffle_or_unshuffle_once` three times with 2 helpers shuffling the shares each time in the opposite order to shuffle.
-/// Order of calling `shuffle_or_unshuffle_once` is shuffle with (H1, H2), (H3, H1) and (H2, H3)
-/// ![Unshuffle steps][unshuffle]
-pub async fn unshuffle_object_shares<F, S, C, I>(
-    input: Vec<I>,
-    random_permutations: (&[u32], &[u32]),
-    ctx: C,
-) -> Result<Vec<I>, Error>
-where
-    C: Context<F, Share = S> + Send,
-    F: Field,
-    I: Resharable<F, Share = S>,
-    S: SecretSharing<F>,
-{
-    let input = shuffle_or_unshuffle_once(
-        input,
-        random_permutations,
-        ShuffleOrUnshuffle::Unshuffle,
-        &ctx,
-        Step3,
-    )
-    .await?;
-    let input = shuffle_or_unshuffle_once(
-        input,
-        random_permutations,
-        ShuffleOrUnshuffle::Unshuffle,
-        &ctx,
-        Step2,
-    )
-    .await?;
-    shuffle_or_unshuffle_once(
-        input,
-        random_permutations,
-        ShuffleOrUnshuffle::Unshuffle,
-        &ctx,
-        Step1,
     )
     .await
 }
@@ -199,7 +158,7 @@ mod tests {
         use crate::ff::Fp31;
         use crate::protocol::context::Context;
         use crate::protocol::sort::shuffle::{
-            get_two_of_three_random_permutations, shuffle_shares, unshuffle_shares,
+            get_two_of_three_random_permutations, shuffle_shares,
         };
         use crate::protocol::QueryId;
         use crate::test_fixture::{Reconstruct, Runner, TestWorld};
@@ -243,50 +202,13 @@ mod tests {
             // Shuffled output should have same inputs
             assert_eq!(hashed_output_secret, hashed_input);
         }
-
-        #[tokio::test]
-        async fn shuffle_unshuffle() {
-            const BATCHSIZE: usize = 5;
-
-            let world = TestWorld::new(QueryId);
-            let input: Vec<u128> = (0..u128::try_from(BATCHSIZE).unwrap()).collect();
-
-            let result = world
-                .semi_honest(
-                    input.clone().into_iter().map(Fp31::from),
-                    |ctx, m_shares| async move {
-                        let perms = get_two_of_three_random_permutations(
-                            BATCHSIZE.try_into().unwrap(),
-                            ctx.prss_rng(),
-                        );
-                        let shuffled = shuffle_shares(
-                            m_shares,
-                            (perms.0.as_slice(), perms.1.as_slice()),
-                            ctx.clone(),
-                        )
-                        .await
-                        .unwrap();
-
-                        unshuffle_shares(
-                            shuffled,
-                            (perms.0.as_slice(), perms.1.as_slice()),
-                            ctx.narrow("unshuffle"),
-                        )
-                        .await
-                        .unwrap()
-                    },
-                )
-                .await;
-
-            assert_eq!(&input[..], &result.reconstruct());
-        }
     }
 
     mod malicious {
         use crate::ff::Fp31;
         use crate::protocol::context::Context;
         use crate::protocol::sort::shuffle::{
-            get_two_of_three_random_permutations, shuffle_shares, unshuffle_shares,
+            get_two_of_three_random_permutations, shuffle_shares,
         };
         use crate::protocol::QueryId;
         use crate::test_fixture::{Reconstruct, Runner, TestWorld};
@@ -331,43 +253,6 @@ mod tests {
 
             // Shuffled output should have same inputs
             assert_eq!(hashed_output_secret, hashed_input);
-        }
-
-        #[tokio::test]
-        async fn shuffle_unshuffle() {
-            const BATCHSIZE: usize = 5;
-
-            let world = TestWorld::new(QueryId);
-            let input: Vec<u128> = (0..u128::try_from(BATCHSIZE).unwrap()).collect();
-
-            let result = world
-                .malicious(
-                    input.clone().into_iter().map(Fp31::from),
-                    |ctx, m_shares| async move {
-                        let perms = get_two_of_three_random_permutations(
-                            BATCHSIZE.try_into().unwrap(),
-                            ctx.prss_rng(),
-                        );
-                        let shuffled = shuffle_shares(
-                            m_shares,
-                            (perms.0.as_slice(), perms.1.as_slice()),
-                            ctx.clone(),
-                        )
-                        .await
-                        .unwrap();
-
-                        unshuffle_shares(
-                            shuffled,
-                            (perms.0.as_slice(), perms.1.as_slice()),
-                            ctx.narrow("unshuffle"),
-                        )
-                        .await
-                        .unwrap()
-                    },
-                )
-                .await;
-
-            assert_eq!(&input[..], &result.reconstruct());
         }
     }
 }
